+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 4.1
 * Fix deeply nested descriptions
 * Fix support for arrays in the path parameters so that they are recursed through as well
+* Improve error handling for invalid spec $refs.